+++ src/red_alert/commands_handler/start_listen_command.rs
@@ -4,7 +4,6 @@ use serenity::model::prelude::ChannelId;
 use serenity::model::prelude::Mention;
 use serenity::prelude::{Context, Mentionable};
 use std::collections::HashMap;
-use std::ops::DerefMut;
 use std::str::FromStr;
 use std::sync::Arc;
 use tokio::sync::RwLock;
@@ -12,33 +11,6 @@ use tokio::sync::RwLock;
 pub(super) struct StartListenRedAlertCommand {
     pub(super) guilds_voices_receivers: Arc<RwLock<HashMap<GuildId, VoiceReceiver>>>,
     pub(super) l10n: L10n,
-}
-
-pub enum StartListenError {
-    SongbirdMissing,
-    ConnectingError,
-}
-
-pub async fn start_listen(
-    guilds_voices_receivers: Arc<RwLock<HashMap<GuildId, VoiceReceiver>>>,
-    ctx: &Context,
-    guild_id: GuildId,
-    channel_id: ChannelId,
-) -> Result<(), StartListenError> {
-    let Some(manager) = songbird::get(ctx).await else {
-        return Err(StartListenError::SongbirdMissing);
-    };
-    let (handler_lock, connection_result) = manager.join(guild_id, channel_id).await;
-    if !connection_result.is_ok() {
-        return Err(StartListenError::ConnectingError);
-    }
-    let mut handler = handler_lock.lock().await;
-    _ = handler.mute(true).await;
-    let voice_receiver = VoiceReceiver::with_configuration(Default::default());
-    voice_receiver.subscribe(handler.deref_mut());
-    let mut guilds_voices_receivers = guilds_voices_receivers.write().await;
-    guilds_voices_receivers.insert(guild_id, voice_receiver);
-    Ok(())
 }
 
 #[async_trait]
