+++ src/red_alert/mod.rs
@@ -2,13 +2,19 @@ mod actions_history;
 mod commands_handler;
 mod guilds_voice_config;
 mod handler;
+mod listen_actions;
+mod monitoring_performer;
+mod recognizer_performer;
 mod voice_config;
 
-pub use actions_history::*;
+use actions_history::*;
 pub use commands_handler::*;
-pub use guilds_voice_config::*;
+use guilds_voice_config::*;
 pub use handler::*;
-pub use voice_config::*;
+use listen_actions::*;
+use monitoring_performer::*;
+use recognizer_performer::*;
+use voice_config::*;
 
 pub(super) const NEW_LINE: &'static str = "\n";
 pub(super) const SPACE: &'static str = " ";
